backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "flowpilot-backend"
version = "0.1.0"
description = "FlowPilot backend services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "bcrypt>=4.2.0",
  "celery>=5.4.0",
  "croniter>=2.0.0",
  "fastapi>=0.115.0",
  "google-api-python-client>=2.140.0",
  "google-auth>=2.35.0",
  "google-auth-oauthlib>=1.2.0",
  "httpx>=0.28.0",
  "openai>=1.50.0",
  "pydantic-settings>=2.6.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.2.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
  "mypy>=1.13.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
src = ["app"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
